# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales Outreach Field Notes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses
      parameters:
        - name: callType
          in: query
          required: false
          schema:
            type: string
            enum: [walk_in, cold_call]
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /visits/{id}/notes:
    get:
      operationId: listNotesForVisit
      tags: [notes]
      summary: List notes for a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note for a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /visits/{id}/media:
    get:
      operationId: listMediaForVisit
      tags: [media]
      summary: List media for a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of media
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload media for a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [image, voice_note, interview, document]
                caption:
                  type: string
              required:
                - file
                - type
      responses:
        "201":
          description: Uploaded media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
  /media/{id}/transcription:
    get:
      operationId: getMediaTranscription
      tags: [media]
      summary: Get transcription status and result for a media item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transcription state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "404":
          description: Not found
    post:
      operationId: requestMediaTranscription
      tags: [media]
      summary: Request (or retry) transcription for a voice media item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Transcription queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "404":
          description: Not found
  /media/{id}:
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List activity events (consumed by Hermes)
      parameters:
        - name: since
          in: query
          required: false
          description: Only events created after this timestamp
          schema:
            type: string
            format: date-time
        - name: type
          in: query
          required: false
          description: Filter by event type prefix (e.g. "visit.")
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Events, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Record a client-side activity event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        sector:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        notes:
          type: string
        mapsUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        priority:
          type: string
          enum: [high, medium, low]
        status:
          type: string
          enum: [not_contacted, contacted, follow_up, converted, not_interested]
        callType:
          type: string
          enum: [walk_in, cold_call]
        routeDay:
          type: integer
        isBonus:
          type: boolean
        buildingGroup:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sector
        - status
        - priority
        - isBonus
        - createdAt
        - updatedAt
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        sector:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        routeDay:
          type: integer
        isBonus:
          type: boolean
        buildingGroup:
          type: string
        notes:
          type: string
        mapsUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        priority:
          type: string
          enum: [high, medium, low]
        status:
          type: string
          enum: [not_contacted, contacted, follow_up, converted, not_interested]
        callType:
          type: string
          enum: [walk_in, cold_call]
      required:
        - name
        - sector
...
    Note:
      type: object
      properties:
        id:
          type: integer
        visitId:
          type: integer
        type:
          type: string
          enum: [text, voice]
        content:
          type: string
        audioUrl:
          type: string
        durationSeconds:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - visitId
        - type
        - createdAt
    CreateNoteBody:
      type: object
      properties:
        type:
          type: string
          enum: [text, voice]
        content:
          type: string
        audioUrl:
          type: string
        durationSeconds:
          type: integer
      required:
        - type
    Media:
      type: object
      properties:
        id:
          type: integer
        visitId:
          type: integer
        type:
          type: string
          enum: [image, voice_note, interview, document]
        url:
          type: string
        filename:
          type: string
        caption:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        transcript:
          type: string
        transcriptionStatus:
          type: string
          enum: [none, pending, processing, done, error]
        transcriptionError:
          type: string
        aiStructured:
          $ref: "#/components/schemas/StructuredNote"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - visitId
        - type
        - url
        - filename
        - transcriptionStatus
        - createdAt
    StructuredNote:
      type: object
      properties:
        summary:
          type: string
        interestLevel:
          type: string
          enum: [hot, warm, cool, cold, unknown]
        objections:
          type: array
          items:
            type: string
        followUpItems:
          type: array
          items:
            type: string
        contactInfo:
          type: string
        nextStep:
          type: string
    Event:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        businessId:
          type: integer
        visitId:
          type: integer
        payload:
          type: object
          additionalProperties: true
        source:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - source
        - createdAt
    CreateEventBody:
      type: object
      properties:
        type:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        businessId:
          type: integer
        visitId:
          type: integer
        payload:
          type: object
          additionalProperties: true
      required:
        - type
...