# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Outreach Field Notes API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessCallType } from "./businessCallType";
import type { BusinessPriority } from "./businessPriority";
import type { BusinessStatus } from "./businessStatus";

export interface Business {
  id: number;
  name: string;
  address?: string;
  phone?: string;
  website?: string;
  sector: string;
  rating?: number;
  reviewCount?: number;
  notes?: string;
  mapsUrl?: string;
  latitude?: number;
  longitude?: number;
  priority: BusinessPriority;
  status: BusinessStatus;
  callType?: BusinessCallType;
  routeDay?: number;
  isBonus: boolean;
  buildingGroup?: string;
  createdAt: Date;
  updatedAt: Date;
}