# lib/api-zod/src/generated/types/createEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Outreach Field Notes API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEventBodyPayload } from "./createEventBodyPayload";

export interface CreateEventBody {
  type: string;
  entityType?: string;
  entityId?: number;
  businessId?: number;
  visitId?: number;
  payload?: CreateEventBodyPayload;
}