# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Outreach Field Notes API
 * OpenAPI spec version: 0.1.0
 */
import type { EventPayload } from "./eventPayload";

export interface Event {
  id: number;
  type: string;
  entityType?: string;
  entityId?: number;
  businessId?: number;
  visitId?: number;
  payload?: EventPayload;
  source: string;
  createdAt: Date;
}